delete working properly, refactoring
