+++ src/main/resources/application.properties
@@ -2,7 +2,7 @@ spring.datasource.driver-class-name=org.postgresql.Driver
 spring.datasource.url=jdbc:postgresql://${WDS_DB_HOST:localhost}:${WDS_DB_PORT:5432}/wds
 spring.datasource.username=${WDS_DB_USER:wds}
 spring.datasource.password=${WDS_DB_PASSWORD:wds}
-
+server.error.include-stacktrace=never
 
 spring.liquibase.change-log=classpath:db/changelog.xml
 spring.liquibase.enabled=true
